backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
Pillow==11.1.0
numpy==2.2.3
scikit-learn==1.6.1
transformers==4.49.0
torch==2.6.0
requests==2.32.3
python-dotenv==1.0.1
